Refactored CRUD operations for the Category API to improve structure and error handling
